refactor(client-runtime): share voice supervisor repository
