+++ packages/client-runtime/package.json
@@ -39,6 +39,10 @@
       "types": "./src/operations/threadTasks.ts",
       "default": "./src/operations/threadTasks.ts"
     },
+    "./operations/voice-supervisor-repository": {
+      "types": "./src/operations/voiceSupervisorRepository.ts",
+      "default": "./src/operations/voiceSupervisorRepository.ts"
+    },
     "./operations/voice-supervisor-tools": {
       "types": "./src/operations/voiceSupervisorTools.ts",
       "default": "./src/operations/voiceSupervisorTools.ts"
