+++ packages/client-runtime/src/operations/index.ts
@@ -2,4 +2,5 @@ export * from "./commands.ts";
 export * from "./projects.ts";
 export * from "./threadSupervisor.ts";
 export * from "./threadTasks.ts";
+export * from "./voiceSupervisorRepository.ts";
 export * from "./voiceSupervisorTools.ts";
